Refactor modular maintenance, search, and remito flows
