+++ frontend/js/__tests__/main.test.js
@@ -211,7 +211,6 @@ describe('manejo de la vista de remito', () => {
     let handleGenerarRemitoClick;
     let handleFinalizarRemitoClick;
     let setLastSavedReportData;
-    let showView;
     let originalAlert;
     let crearRemitoMock;
 
@@ -268,7 +267,6 @@ describe('manejo de la vista de remito', () => {
         handleGenerarRemitoClick = mainModule.__testables__.handleGenerarRemitoClick;
         handleFinalizarRemitoClick = mainModule.__testables__.handleFinalizarRemitoClick;
         setLastSavedReportData = mainModule.__testables__.setLastSavedReportDataForTests;
-        showView = mainModule.__testables__.showView;
 
         originalAlert = window.alert;
         window.alert = jest.fn();
@@ -330,11 +328,6 @@ describe('manejo de la vista de remito', () => {
 
         const formView = document.getElementById('tab-nuevo');
         const remitoView = document.getElementById('remito-servicio');
-        expect(formView.classList.contains('hidden')).toBe(false);
-        expect(remitoView.classList.contains('hidden')).toBe(true);
-
-        showView('remito-servicio');
-
         expect(formView.classList.contains('hidden')).toBe(true);
         expect(remitoView.classList.contains('hidden')).toBe(false);
